separate scorers into separate files to make it possible to import one without depending on the others
