+++ asapdiscovery-workflows/asapdiscovery/workflows/docking_workflows/small_scale_docking.py
@@ -39,10 +39,10 @@ from asapdiscovery.docking.docking_data_validation import DockingResultCols
 from asapdiscovery.docking.openeye import POSITDocker
 from asapdiscovery.docking.scorer import (
     ChemGauss4Scorer,
-    FINTScorer,
-    MetaScorer,
-    MLModelScorer,
 )
+from asapdiscovery.docking.fint_scorer import FINTScorer
+from asapdiscovery.docking.ml_scorer import MLModelScorer
+from asapdiscovery.docking.meta_scorer import MetaScorer
 from asapdiscovery.ml.models import ASAPMLModelRegistry
 from asapdiscovery.modeling.protein_prep import ProteinPrepper
 from asapdiscovery.simulation.simulate import OpenMMPlatform, VanillaMDSimulator
